requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.31.0
nltk==3.8.1
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
python-dateutil==2.8.2
pyyaml==6.0.1
sqlalchemy==2.0.25
bcrypt==4.1.2
pyjwt==2.8.0
langchain==0.1.4
transformers==4.36.2 
markdown2==2.4.10
SQLAlchemy==2.0.25
Werkzeug==3.1.3 